rewrite llama
