--- python/llm/src/ipex_llm/transformers/models/llama32.py
@@ -1,247 +1,0 @@
-#
-# Copyright 2016 The BigDL Authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# Some parts of this file is adapted from
-# https://github.com/huggingface/transformers/blob/main/src/transformers/models/llama/modeling_llama.py
-# which is licensed under Apache License 2.0:
-#
-# Copyright 2022 EleutherAI and the HuggingFace Inc. team. All rights reserved.
-#
-# This code is based on EleutherAI's GPT-NeoX library and the GPT-NeoX
-# and OPT implementations in this library. It has been modified from its
-# original forms to accommodate minor architectural differences compared
-# to GPT-NeoX and OPT used by the Meta AI team that trained the model.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import math
-import torch
-
-from typing import Optional, Tuple, Union
-from transformers.cache_utils import Cache
-from transformers.modeling_outputs import BaseModelOutputWithPast
-from transformers.models.llama.modeling_llama import apply_rotary_pos_emb
-
-from ipex_llm.utils.common import invalidInputError
-from ipex_llm.transformers.models.common import scaled_dot_product_attention
-from ipex_llm.transformers.models.utils import should_use_fuse_rope
-from ipex_llm.transformers.models.utils import use_quantize_kv_cache
-from ipex_llm.transformers.models.utils import should_use_compresskv, \
-    is_enough_kv_cache_room_4_36
-from ipex_llm.transformers.kv import DynamicNormalCache, DynamicFp8Cache, DynamicCompressCache, \
-    DynamicCompressFp8Cache
-
-
-def llama_model_forward(
-    self,
-    input_ids: torch.LongTensor = None,
-    attention_mask: Optional[torch.Tensor] = None,
-    position_ids: Optional[torch.LongTensor] = None,
-    past_key_values: Optional[Cache] = None,
-    inputs_embeds: Optional[torch.FloatTensor] = None,
-    use_cache: Optional[bool] = None,
-    output_attentions: Optional[bool] = None,
-    output_hidden_states: Optional[bool] = None,
-    return_dict: Optional[bool] = None,
-    cache_position: Optional[torch.LongTensor] = None,
-) -> Union[Tuple, BaseModelOutputWithPast]:
-    output_attentions = (
-        output_attentions if output_attentions is not None
-        else self.config.output_attentions
-    )
-    output_hidden_states = (
-        output_hidden_states if output_hidden_states is not None
-        else self.config.output_hidden_states
-    )
-    use_cache = use_cache if use_cache is not None else self.config.use_cache
-
-    # IPEX-LLM OPT start: kv cache and quantize kv cache
-    inputs = input_ids if input_ids is not None else inputs_embeds
-    use_cache = True if inputs.device.type == "xpu" else use_cache
-    use_quantize_kv = use_quantize_kv_cache(
-        self.layers[0].mlp.down_proj, inputs,
-        self.config.num_attention_heads // self.config.num_key_value_heads
-    )
-    use_compresskv = should_use_compresskv(inputs, inputs.shape[1]) or \
-        isinstance(past_key_values, DynamicCompressCache)
-    # disable llama3.2 1b for prefill performance and output quality
-    use_compresskv = use_compresskv and self.config.hidden_size != 2048
-    if use_cache:
-        if use_compresskv and not isinstance(past_key_values, DynamicCompressCache):
-            if use_quantize_kv:
-                past_key_values = DynamicCompressFp8Cache.from_legacy_cache(past_key_values)
-            else:
-                past_key_values = DynamicCompressCache.from_legacy_cache(past_key_values)
-        elif use_quantize_kv and not isinstance(past_key_values, DynamicFp8Cache):
-            past_key_values = DynamicFp8Cache.from_legacy_cache(past_key_values)
-        elif (
-            not use_quantize_kv
-            and not use_compresskv
-            and not isinstance(past_key_values, DynamicNormalCache)
-        ):
-            past_key_values = DynamicNormalCache.from_legacy_cache(past_key_values)
-
-    # IPEX-LLM OPT end
-
-    return_dict = return_dict if return_dict is not None else self.config.use_return_dict
-
-    invalidInputError((input_ids is None) ^ (inputs_embeds is None),
-                      "You cannot specify both input_ids and inputs_embeds at the same time, "
-                      "and must specify either one")
-
-    if inputs_embeds is None:
-        inputs_embeds = self.embed_tokens(input_ids)
-
-    if cache_position is None:
-        past_seen_tokens = past_key_values.get_seq_length() if past_key_values is not None else 0
-        cache_position = torch.arange(
-            past_seen_tokens, past_seen_tokens + inputs_embeds.shape[1], device=inputs_embeds.device
-        )
-    if position_ids is None:
-        position_ids = cache_position.unsqueeze(0)
-
-    causal_mask = self._update_causal_mask(
-        attention_mask, inputs_embeds, cache_position, past_key_values, output_attentions
-    )
-    hidden_states = inputs_embeds
-
-    # create position embeddings to be shared across the decoder layers
-    position_embeddings = self.rotary_emb(hidden_states, position_ids)
-
-    # IPEX-LLM OPT start: use fused rope
-    if (should_use_fuse_rope(hidden_states, position_ids, False)
-            and self.rotary_emb.rope_type in ["default", "llama3"]):
-        position_embeddings = self.rotary_emb.inv_freq
-    # IEPX_LLM OPT end
-
-    # decoder layers
-    all_hidden_states = () if output_hidden_states else None
-    all_self_attns = () if output_attentions else None
-    next_decoder_cache = None
-
-    for decoder_layer in self.layers:
-        if output_hidden_states:
-            all_hidden_states += (hidden_states,)
-
-        layer_outputs = decoder_layer(
-            hidden_states,
-            attention_mask=causal_mask,
-            position_ids=position_ids,
-            past_key_value=past_key_values,
-            output_attentions=output_attentions,
-            use_cache=use_cache,
-            cache_position=cache_position,
-            position_embeddings=position_embeddings,
-        )
-
-        hidden_states = layer_outputs[0]
-
-        if use_cache:
-            next_decoder_cache = layer_outputs[2 if output_attentions else 1]
-
-        if output_attentions:
-            all_self_attns += (layer_outputs[1],)
-
-    hidden_states = self.norm(hidden_states)
-
-    # add hidden states from the last decoder layer
-    if output_hidden_states:
-        all_hidden_states += (hidden_states,)
-
-    next_cache = next_decoder_cache if use_cache else None
-
-    if not return_dict:
-        return tuple(v for v in [hidden_states, next_cache, all_hidden_states, all_self_attns]
-                     if v is not None)
-    return BaseModelOutputWithPast(
-        last_hidden_state=hidden_states,
-        past_key_values=next_cache,
-        hidden_states=all_hidden_states,
-        attentions=all_self_attns,
-    )
-
-
-def llama_attention_forward(
-    self,
-    hidden_states: torch.Tensor,
-    attention_mask: Optional[torch.Tensor] = None,
-    position_ids: Optional[torch.LongTensor] = None,
-    past_key_value: Optional[Cache] = None,
-    output_attentions: bool = False,
-    use_cache: bool = False,
-    cache_position: Optional[torch.LongTensor] = None,
-    position_embeddings: Optional[Tuple[torch.Tensor, torch.Tensor]]=None,
-    **kwargs,
-) -> Tuple[torch.Tensor, Optional[torch.Tensor], Optional[Tuple[torch.Tensor]]]:
-    bsz, q_len, _ = hidden_states.size()
-
-    # [CompressKV]
-    use_compresskv = isinstance(past_key_value, DynamicCompressCache)
-
-    qkv = self.qkv_proj(hidden_states)
-    qkv = qkv.view(bsz, q_len, self.num_heads + 2 * self.num_key_value_heads, self.head_dim)
-    qkv = qkv.transpose(1, 2)
-    query_states, key_states, value_states = qkv.split([self.num_heads,
-                                                        self.num_key_value_heads,
-                                                        self.num_key_value_heads], dim=1)
-
-    if isinstance(position_embeddings, torch.Tensor):
-        import xe_addons
-        inv_freq = position_embeddings
-        xe_addons.rotary_half_inplaced(inv_freq, position_ids, query_states, key_states)
-    else:
-        if position_embeddings is None:
-            cos, sin = self.rotary_emb(value_states, position_ids)
-        else:
-            cos, sin = position_embeddings
-        query_states, key_states = apply_rotary_pos_emb(query_states, key_states, cos, sin)
-
-    if past_key_value is not None:
-        # [CompressKV]
-        if use_compresskv:
-            enough_kv_room = is_enough_kv_cache_room_4_36(past_key_value, self.layer_idx,
-                                                          q_len)
-            key_states, value_states = past_key_value.update(
-                key_states, value_states, self.layer_idx,
-                query_states, attention_mask, self.num_key_value_groups,
-                self.config, enough_kv_room, 256)
-        else:
-            key_states, value_states = past_key_value.update(key_states, value_states,
-                                                             self.layer_idx, None)
-
-    attn_weights = None
-    attn_output = scaled_dot_product_attention(
-        query_states, key_states, value_states,
-        attention_mask, q_len == key_states.size(2)
-    )
-
-    attn_output = attn_output.transpose(1, 2).contiguous()
-    attn_output = attn_output.reshape(bsz, q_len, -1)
-    attn_output = self.o_proj(attn_output)
-
-    if not output_attentions:
-        attn_weights = None
-
-    return attn_output, attn_weights, past_key_value
